LayoutTests/fast/canvas/webgl/bug-32456.html

<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css"/>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>

description('Regression test for <a href="https://bugs.webkit.org/show_bug.cgi?id=32456">https://bugs.webkit.org/show_bug.cgi?id=32456</a> : <code>Errors in WebGL[T]Array set() and get() methods</code>');

debug("Testing WebGLFloatArray get / set");
var array = new WebGLFloatArray([1.5]);
shouldBe("array.get(0)", "1.5");
shouldBe("array[0]", "1.5");
shouldBeUndefined("array.set(0, 2.5)");
shouldBe("array.get(0)", "2.5");
shouldBe("array[0]", "2.5");

debug("Testing WebGLByteArray get / set");
array = new WebGLByteArray(1);
shouldBeUndefined("array.set(0, 127)");
shouldBe("array.get(0)", "127");
shouldBe("array[0]", "127");
shouldBeUndefined("array.set(0, -128)");
shouldBe("array.get(0)", "-128");
shouldBe("array[0]", "-128");

debug("Testing WebGLUnsignedByteArray get / set");
array = new WebGLUnsignedByteArray(1);
shouldBeUndefined("array.set(0, 255)");
shouldBe("array.get(0)", "255");
shouldBe("array[0]", "255");

debug("Testing WebGLShortArray get / set");
array = new WebGLShortArray(1);
shouldBeUndefined("array.set(0, 32767)");
shouldBe("array.get(0)", "32767");
shouldBe("array[0]", "32767");
shouldBeUndefined("array.set(0, -32768)");
shouldBe("array.get(0)", "-32768");
shouldBe("array[0]", "-32768");

debug("Testing WebGLUnsignedShortArray get / set");
array = new WebGLUnsignedShortArray(1);
shouldBeUndefined("array.set(0, 65535)");
shouldBe("array.get(0)", "65535");
shouldBe("array[0]", "65535");

debug("Testing WebGLIntArray get / set");
array = new WebGLIntArray(1);
shouldBeUndefined("array.set(0, -2147483648)");
shouldBe("array.get(0)", "-2147483648");
shouldBe("array[0]", "-2147483648");
shouldBeUndefined("array.set(0, 2147483647)");
shouldBe("array.get(0)", "2147483647");
shouldBe("array[0]", "2147483647");

debug("Testing WebGLUnsignedIntArray get / set");
array = new WebGLUnsignedIntArray(1);
shouldBeUndefined("array.set(0, 4294967295)");
shouldBe("array.get(0)", "4294967295");
shouldBe("array[0]", "4294967295");

successfullyParsed = true;

</script>
<script src="../../js/resources/js-test-post.js"></script>

</body>
</html>
