LayoutTests/fast/canvas/webgl/bug-32456-expected.txt

Regression test for https://bugs.webkit.org/show_bug.cgi?id=32456 : Errors in WebGL[T]Array set() and get() methods

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing WebGLFloatArray get / set
PASS array.get(0) is 1.5
PASS array[0] is 1.5
PASS array.set(0, 2.5) is undefined.
PASS array.get(0) is 2.5
PASS array[0] is 2.5
Testing WebGLByteArray get / set
PASS array.set(0, 127) is undefined.
PASS array.get(0) is 127
PASS array[0] is 127
PASS array.set(0, -128) is undefined.
PASS array.get(0) is -128
PASS array[0] is -128
Testing WebGLUnsignedByteArray get / set
PASS array.set(0, 255) is undefined.
PASS array.get(0) is 255
PASS array[0] is 255
Testing WebGLShortArray get / set
PASS array.set(0, 32767) is undefined.
PASS array.get(0) is 32767
PASS array[0] is 32767
PASS array.set(0, -32768) is undefined.
PASS array.get(0) is -32768
PASS array[0] is -32768
Testing WebGLUnsignedShortArray get / set
PASS array.set(0, 65535) is undefined.
PASS array.get(0) is 65535
PASS array[0] is 65535
Testing WebGLIntArray get / set
PASS array.set(0, -2147483648) is undefined.
PASS array.get(0) is -2147483648
PASS array[0] is -2147483648
PASS array.set(0, 2147483647) is undefined.
PASS array.get(0) is 2147483647
PASS array[0] is 2147483647
Testing WebGLUnsignedIntArray get / set
PASS array.set(0, 4294967295) is undefined.
PASS array.get(0) is 4294967295
PASS array[0] is 4294967295
PASS successfullyParsed is true

TEST COMPLETE
